move debug-info mixin into DS.Model directly
